+++ packages/shared/src/data/config.ts
@@ -15,5 +15,5 @@ export default {
   defaultDuration: 310,
   minChunkSize: 5295308,
   loggerLevel: 1 as LoggerLevel,
-  version: "2.5.0",
+  version: "3.0.0",
 } satisfies ConfigSchema;