+++ changelog.md
@@ -16,6 +16,7 @@ This release contains **many breaking changes**, including a complete rewrite of
 - [!] Drop built-in typebox types
 - Added support for Wistia Embed (`https://fast.wistia.net/embed/iframe/XXX`)
 - Fixed Vimeo embed with `app_id` query param
+- Fixed missing `extraOpts` field in retry request `translateVideo` after AUDIO_REQUESTED
 - Improved raw `.mp3` / `.webm` link matching (fixed [voice-over-translation#1384](https://github.com/ilyhalight/voice-over-translation/issues/1384))
 
 ## Ext