+++ changelog.md
@@ -16,7 +16,7 @@ This release contains **many breaking changes**, including a complete rewrite of
 - [!] Drop `vot-backend` provider support. Now, lib will try to translate `.m3u8` files directly
 - [!] Drop `incestflix` service support
 - [!] Reworked logic of `VOTClient` class. Now, all logic based in separated `Providers` and you can easy add your own provider for any service
-- [!] Removed `VOTWorker` class. Use `VOTClient` with `provider: VOTWorkerProvider` instead
+- [!] Removed `VOTWorkerClient` class. Use `VOTClient` with `provider: VOTWorkerProvider` instead
 - [!] Removed `MinimalClient` class
 - [!] Now, only `translateVideo`, `translateStream` and `getSubtitles` available as public methods in `VOTClient`. All other specific methods moved to `Providers` and you can use them directly from provider instance, e.g. `client.provider.translateVideoCache(...)`
 - [!] Drop built-in typebox types